refactor(settings-dialog): split nav into hook + sidebar + mobile picker
Was 339 lines doing nav config + hash/localStorage sync + arrow-key
handler + sidebar render + mobile picker + dialog shell. Splits into:

- settings-nav-config.ts -- grouped + flat nav definitions and the
  STORAGE_KEY constant. Pure data, no React.
- use-settings-nav.ts -- hook owning activeSection, keyboardSelectedIndex,
  hash/localStorage round-trip, focus-first-button, and the window
  keydown listener. Returns { activeSection, setActiveSection,
  keyboardSelectedIndex }.
- settings-nav-sidebar.tsx -- desktop sidebar with the grouped nav.
- settings-nav-mobile.tsx -- below-md Select picker.
- settings-dialog.tsx (104 lines) -- dialog shell composing the parts.

No behavior change. The hook is intentionally "always on" while
`open` is true, mirroring the original component's event-listener
lifecycle exactly.
